Remove arguments parameter to GET request
I'm not sure why this is happening, but `arguments` doesn't seem to have any effect in the API call. The same data is returned with or without it.

Additionally, the file currently does not run as is, there's an error:

```
Traceback (most recent call last):
  File "demo.py", line 181, in <module>
    id_list = board['lists'][0]['id']
KeyError: 'lists'
```

I believe this is because lists default to `none`, and since the data variable does not seem to have any effect, no lists are returned. Why data has no effect, I'm not sure. Lumping the lists into the params variable fixes this problem, and the file runs without error.